adjust the mock states mempool
