+++ python/sgl_jax/test/layers/test_lightning_backend_dp.py
@@ -17,12 +17,10 @@ import pytest
 from jax.sharding import NamedSharding
 from jax.sharding import PartitionSpec as P
 
-from sgl_jax.srt.layers.attention.hybrid_linear_attn_backend import (
-    MockRecurrentStatePool,
-)
 from sgl_jax.srt.layers.attention.linear.lightning_backend import LightningAttnBackend
 from sgl_jax.srt.model_executor.forward_batch_info import ForwardMode
 from sgl_jax.srt.utils.mesh_utils import create_device_mesh
+from sgl_jax.test.layers.mock_recurrent_state_pool import MockRecurrentStatePool
 
 try:
     from sgl_jax.srt.kernels.simple_gla.simple_gla import fused_recurrent_simple_gla
