+++ python/sgl_jax/test/test_lightning_backend.py
@@ -18,13 +18,11 @@ import jax.numpy as jnp
 import numpy as np
 import pytest
 
-from sgl_jax.srt.layers.attention.hybrid_linear_attn_backend import (
-    MockRecurrentStatePool,
-)
 from sgl_jax.srt.layers.attention.linear.lightning_backend import LightningAttnBackend
 from sgl_jax.srt.model_executor.forward_batch_info import ForwardMode
 from sgl_jax.srt.utils.mesh_utils import create_device_mesh
 from sgl_jax.test.gla_reference import naive_gla_decode, naive_gla_prefill
+from sgl_jax.test.layers.mock_recurrent_state_pool import MockRecurrentStatePool
 
 mesh = create_device_mesh(ici_parallelism=[1, -1], dcn_parallelism=[1, 1])
